login system

+++ middlewares/users/userValidators.js
@@ -58,7 +58,7 @@ const addUserValidationHandler = function (req, res, next) {
     next();
   } else {
     // remove uploaded files
-    if (req.files.length > 0) {
+    if (req && req.files && req.files.length > 0) {
       const { filename } = req.files[0];
       unlink(
         path.join(__dirname, `/../public/uploads/avatars/${filename}`),
